+++ brender/src/test/java/com/sefford/brender/adapters/RendererAdapterTest.java
@@ -28,7 +28,8 @@ import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
-import org.robolectric.RobolectricTestRunner;
+import org.robolectric.RobolectricGradleTestRunner;
+import org.robolectric.RobolectricGradleTestRunner;
 import org.robolectric.RuntimeEnvironment;
 
 import static org.junit.Assert.*;
@@ -40,7 +41,7 @@ import static org.mockito.MockitoAnnotations.initMocks;
  *
  * @author Saul Diaz <sefford@gmail.com>
  */
-@RunWith(RobolectricTestRunner.class)
+@RunWith(RobolectricGradleTestRunner.class)
 public class RendererAdapterTest {
 
     static final int EXPECTED_COUNT = 1;