+++ brender/src/test/java/com/sefford/brender/filters/NullFilterTest.java
@@ -18,12 +18,12 @@ package com.sefford.brender.filters;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.robolectric.RobolectricTestRunner;
+import org.robolectric.RobolectricGradleTestRunner;
 
 import static org.junit.Assert.assertNull;
 import static org.mockito.MockitoAnnotations.initMocks;
 
-@RunWith(RobolectricTestRunner.class)
+@RunWith(RobolectricGradleTestRunner.class)
 public class NullFilterTest {
 
     private NullFilter filter;